Talk2Databases/requirements.txt
langchain
langchain-openai
mysql-connector-python
streamlit
python-dotenv